+++ cmd/generator/generator.go
@@ -1,7 +1,7 @@
 package main
 
 import (
-	"github.com/superhacker2002/parentheses/internal/controller"
+	"github.com/superhacker2002/parentheses/internal/controller/generator"
 	"log"
 	"math/rand"
 	"net/http"
@@ -9,7 +9,7 @@ import (
 )
 
 func main() {
-	controller.SetRoutes()
+	generator.SetRoutes()
 	rand.Seed(time.Now().UnixNano())
 	log.Fatal(http.ListenAndServe("localhost:8080", nil))
 }
