+++ internal/controller/generator/generator.go
@@ -1,4 +1,4 @@
-package controller
+package generator
 
 import (
 	"fmt"
