+++ internal/controller/generator/generator_test.go
@@ -1,4 +1,4 @@
-package controller
+package generator
 
 import (
 	"github.com/stretchr/testify/assert"
